+++ storage/sqlite/migrations/000028_qualification_offerings.down.sql
@@ -1,0 +1,2 @@
+DROP INDEX IF EXISTS qualification_offerings_profile_idx;
+DROP TABLE IF EXISTS qualification_offerings;